requirements.txt
certifi==2019.3.9
chardet==3.0.4
croniter==0.3.30
idna==2.8
python-crontab==2.3.6
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
slackbot==0.5.3
slacker==0.13.0
urllib3==1.24.3
websocket-client==0.44.0
